requirements.txt
crewai>=0.1.0
facebook-business>=17.0.0
pandas>=1.5.0
python-dotenv>=0.19.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
slack-sdk>=3.19.5
slack-bolt>=1.16.1
APScheduler>=3.9.1
pytz>=2023.3
